+++ app/src/main/java/com/demasu/testpixeldungeon/actors/mobs/Bandit.java
@@ -35,7 +35,7 @@ public class Bandit extends Thief {
         spriteClass = BanditSprite.class;
 
         name = Dungeon.getCurrentDifficulty().mobPrefix() + name;
-        setHT( getHT() * Dungeon.getCurrentDifficulty().mobHPModifier() );
+        setHT( (int) (getHT() * Dungeon.getCurrentDifficulty().mobHPModifier()) );
         setHP( getHT() );
     }
